fixed unwanted revised delay
It was a small change and I technically fixed it the simplest way possible by using the "edit" button that forks and allows edits. I found what the sleep was before the change and manually changed it back. It seemed like overkill to fork, clone, fix, push to private repo and then submit pull request.

It would make sense if there were A LOT of changes but it was a 1 digit difference under commit "increased delay..."